JS overall
Removed the city variable to stop the error where the state would returned from geolocate api instead of the state. The weatherbit api works fine with just the city. I also added a loading message to be displayed while the weather data is being fetched.  On the backend the view & path was updated account for the chaneges. A sketch up of the ui help focus development, see sketch.png.